+++ src/ResponseTypes/BearerTokenResponse.php
@@ -17,6 +17,7 @@ namespace League\OAuth2\Server\ResponseTypes;
 use League\OAuth2\Server\Entities\AccessTokenEntityInterface;
 use LogicException;
 use Psr\Http\Message\ResponseInterface;
+use SensitiveParameter;
 
 use function array_merge;
 use function json_encode;
@@ -75,8 +76,10 @@ class BearerTokenResponse extends AbstractResponseType
      *
      * @return array<array-key,mixed>
      */
-    protected function getExtraParams(AccessTokenEntityInterface $accessToken): array
-    {
+    protected function getExtraParams(
+        #[SensitiveParameter]
+        AccessTokenEntityInterface $accessToken
+    ): array {
         return [];
     }
 }